refactor(U12): move the column-role helpers into @fusion/core; convert live-agent-count
The server side asks the same question the dashboard does — live agent counting gates
on "intake or hold" with the identical legacy-id fallback — so keeping the helper in
the dashboard would have put the guess in two packages. Canonical definition moves to
packages/core/src/column-roles.ts; the dashboard file becomes a re-export so its
call sites keep a local import path.

live-agent-count had the pair inlined TWICE, in a producer and its own defaulting
consumer. They now agree by construction rather than by someone remembering both.

live-agent-count 2 -> 0, archive-planning 1 -> 0 (comment prose).
